+++ src/plugin.cpp
@@ -9,5 +9,5 @@ void init(rack::Plugin *p) {
 	p->addModel(modelCV_Map);
 	p->addModel(modelCV_Pam);
 	p->addModel(modelRotorA);
-	p->addModel(modelRePlay);
+	p->addModel(modelReMoveLight);
 }